Separate dom with index file
